refactor: move useful function into seperate module
