+++ libs/utility_functions/__init__.py
@@ -1,0 +1,18 @@
+"""
+Copyright (c) 2024 MPI-M, Clara Bayley
+
+----- Microphysics Test Cases -----
+File: __init__.py
+Project: utility_functions
+Created Date: Monday 2nd September 2024
+Author: Clara Bayley (CB)
+Additional Contributors:
+-----
+Last Modified: Monday 2nd September 2024
+Modified By: CB
+-----
+License: BSD 3-Clause "New" or "Revised" License
+https://opensource.org/licenses/BSD-3-Clause
+-----
+File Description:
+"""
